kvmtool: assume dead vcpus are paused too
vcpus that have exited their main loop can be treated as paused too when we're
trying to pause the machine to change mappings.

This might happen during shutdown when we're trying to unmap ioports but some
of the vcpus already exited.

In practice, we would hang trying to pause a vcpu that's already exited.

Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>